Initial migration to V2 Energy API (#2474)
* Initial Energy API refactor

* Allow charging in energy storage BEs

* Remove maven for experimental fabric

+++ RebornCore/src/main/java/reborncore/common/blockentity/MachineBaseBlockEntity.java
@@ -327,8 +327,8 @@ public class MachineBaseBlockEntity extends BlockEntity implements BlockEntityTi
 	}
 
 	@Override
-	public double getEuPerTick(double baseEu) {
-		return baseEu * powerMultiplier;
+	public long getEuPerTick(long baseEu) {
+		return (long) (baseEu * powerMultiplier);
 	}
 
 	@Override
